slack class file test
